% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koller-data.R
\name{koller}
\alias{koller}
\title{Responses to the SCC scale and CSII scale (n=1013).}
\format{
A 1013 x 24 data frame.
}
\source{
Koller, Floh, Zauner, Rusch (2013) "Persuasibility and the Self – Investigating Heterogeneity among Consumers". Australasian Marketing Journal, 21, 94-104.
}
\description{
1013 respondents answered the items of the consumer susceptibility top interpersonal influence (CSII) and the self-concept clarity (SCC) scale. The answers are on a five-point-Likert scale with 1 meaning "fully agree" and 5 meaning "fully disagree". Note that on scc the 10th item is reversed coded to align with the other items.
}
