% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{plot3dstatic}
\alias{plot3dstatic}
\title{plot3dstatic: static 3D plots}
\usage{
plot3dstatic(x, plot.dim = c(1, 2, 3), main, xlab, ylab, zlab, col, ...)
}
\arguments{
\item{x}{object}

\item{plot.dim}{dimensions to plot}

\item{main}{main title}

\item{xlab}{label for x axis}

\item{ylab}{label for y axis}

\item{zlab}{label for z axis}

\item{col}{color}

\item{...}{other arguments}
}
\description{
A static 3d plot S3 generic
}
\details{
A static 3d plot
}
