% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R, R/procruster.R
\name{procruster}
\alias{procruster}
\title{procruster: a procrustes function}
\usage{
procruster(x)

procruster(x)
}
\arguments{
\item{x}{numeric matrix}
}
\value{
a matrix

a matrix
}
\description{
procruster: a procrustes function

procruster: a procrustes function
}
