% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ringdat.R
\name{ringdat}
\alias{ringdat}
\title{Two interlocking rings}
\format{
A 500 x 4 matrix.
}
\description{
Artifical data of two interlocking rings in 3D space. The first three columns are the coordinates and the last column is a color designation.
}
