% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smacofxDeleteOne.R
\name{smacofxDeleteOne}
\alias{smacofxDeleteOne}
\title{Helper function to conduct jackknife MDS}
\usage{
smacofxDeleteOne(
  object,
  delta,
  weightmat,
  init,
  ndim,
  type,
  verbose = FALSE,
  itmaxi = 10000
)
}
\arguments{
\item{object}{Object of class smacofP if used as method or another object inheriting from smacofB. Note we assume the MDS model was fitted on a symmetric matrix/data frame or dist object}

\item{delta}{the data (symmetric matrix, data frame or dist object)}

\item{weightmat}{weighting matrix}

\item{init}{starting configuration}

\item{ndim}{target dimension of the mds}

\item{type}{type of MDS}

\item{verbose}{print progress}

\item{itmaxi}{maximum iterations of the MDS procedure}
}
\value{
An array of size n with n coonfigurations
}
\description{
Function deletes every object row and columns once and fits the MDS in object and returns the configuration. The deleted row is set to 0 in the configuration. Is meant for smacofx functions, but should also work for every smacof models.
}
