% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spheredat.R
\name{spheredat}
\alias{spheredat}
\title{Clelia curve on a sphere}
\format{
A 500 x 4 matrix.
}
\description{
Artifical data of data sampled along a Clelia curve on a sphere. The first three columns are the coordinates and the last column is a color designation (viridis palette).
}
