% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spp.R
\name{spp}
\alias{spp}
\title{Calculating stress per point}
\usage{
spp(dhat, confdist, weightmat)
}
\arguments{
\item{dhat}{a dist object or symmetric matrix of dissimilarities}

\item{confdist}{a dist object or symmetric matrix of fitted distances}

\item{weightmat}{dist objetc or symmetric matrix of weights}
}
\value{
a list
}
\description{
Calculating stress per point
}
