% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udat.R
\name{udat}
\alias{udat}
\title{Noisy Data on a U Fold}
\format{
A 500 x 4 matrix.
}
\description{
Artifical data of data on a 3D U fold with noise that increases towards the edges. The first three columns are the coordinates and the last column is a color designation (viridis palette).
}
